pyproject.toml
[project]
name = "yaldabaranth"
version = "0.1.0"
description = "A madman's game."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "esper>=3.4",
    "numpy>=2.3.5",
    "pygame-ce>=2.5.6",
    "raylib>=5.5.0.3",
    "scipy>=1.16.3",
    "tcod>=19.6.0",
]
